pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devrun"
version = "0.1.0"
description = "Modular developer task orchestration system for heterogeneous compute backends."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "typer>=0.9",
    "pydantic>=1.10,<3",
    "pyyaml>=6.0",
    "requests>=2.28",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "ruff>=0.1",
]

[project.scripts]
devrun = "devrun.cli:main"

[tool.setuptools.packages.find]
include = ["devrun*"]
